refactor: only transform visible text search results to display results to avoid unnecessary work since most text search results are always hidden
