+++ packages/renderer-worker/src/parts/ViewletSearch/ViewletSearchHandleUpdate.js
@@ -1,56 +1,9 @@
 import * as ErrorHandling from '../ErrorHandling/ErrorHandling.js'
-import * as IconTheme from '../IconTheme/IconTheme.js'
 import * as ScrollBarFunctions from '../ScrollBarFunctions/ScrollBarFunctions.js'
 import * as TextSearch from '../TextSearch/TextSearch.js'
 import * as TextSearchResultType from '../TextSearchResultType/TextSearchResultType.js'
 import * as Workspace from '../Workspace/Workspace.js'
 import * as ViewletSearchStatusMessage from './ViewletSearchStatusMessage.js'
-
-const toDisplayResults = (results, itemHeight, resultCount, searchTerm) => {
-  // results.sort(compareResults)
-  const displayResults = []
-  let i = -1
-  const setSize = resultCount
-  for (const result of results) {
-    i++
-    switch (result.type) {
-      case TextSearchResultType.File:
-        const path = result.text
-        const absolutePath = Workspace.getAbsolutePath(path)
-        const baseName = Workspace.pathBaseName(path)
-        displayResults.push({
-          title: absolutePath,
-          type: TextSearchResultType.File,
-          text: baseName,
-          icon: IconTheme.getFileIcon({ name: baseName }),
-          posInSet: i + 1,
-          setSize,
-          top: i * itemHeight,
-          lineNumber: result.lineNumber,
-          matchStart: 0,
-          matchLength: 0,
-        })
-        break
-      case TextSearchResultType.Match:
-        displayResults.push({
-          title: result.text,
-          type: TextSearchResultType.Match,
-          text: result.text,
-          icon: '',
-          posInSet: i + 1,
-          setSize,
-          top: i * itemHeight,
-          lineNumber: result.lineNumber,
-          matchStart: result.start,
-          matchLength: searchTerm.length,
-        })
-        break
-      default:
-        break
-    }
-  }
-  return displayResults
-}
 
 const getResultCounts = (results) => {
   let resultCount = 0
@@ -95,9 +48,9 @@ export const handleUpdate = async (state, update) => {
       throw new Error(`results must be of type array`)
     }
     const { fileCount, resultCount } = getResultCounts(results)
-    const displayResults = toDisplayResults(results, itemHeight, resultCount, value)
+    // const displayResults = toDisplayResults(results, itemHeight, resultCount, value)
     const message = ViewletSearchStatusMessage.getStatusMessage(resultCount, fileCount)
-    const total = displayResults.length
+    const total = results.length
     const contentHeight = total * itemHeight
     const listHeight = height - headerHeight
     const scrollBarHeight = ScrollBarFunctions.getScrollBarHeight(height, contentHeight, minimumSliderSize)
@@ -107,7 +60,7 @@ export const handleUpdate = async (state, update) => {
     return {
       ...partialNewState,
       value,
-      items: displayResults,
+      items: results,
       message,
       maxLineY: maxLineY,
       scrollBarHeight,
