+++ packages/renderer-worker/test/ViewletSearch.test.js
@@ -110,16 +110,18 @@ test('setValue - one match in one file', async () => {
     value: 'abc',
     items: [
       {
-        icon: '',
-        text: 'file-1.txt',
-        title: '/file-1.txt',
         type: TextSearchResultType.File,
+        text: './file-1.txt',
+        start: 0,
+        end: 0,
+        lineNumber: 0,
       },
       {
-        icon: '',
-        text: 'abc',
-        title: 'abc',
         type: TextSearchResultType.Match,
+        text: 'abc',
+        start: 0,
+        end: 0,
+        lineNumber: 0,
       },
     ],
     message: 'Found 1 result in 1 file',
@@ -158,22 +160,25 @@ test('setValue - two matches in one file', async () => {
     value: 'abc',
     items: [
       {
-        icon: '',
-        text: 'file-1.txt',
-        title: '/file-1.txt',
         type: TextSearchResultType.File,
+        text: './file-1.txt',
+        start: 0,
+        end: 0,
+        lineNumber: 0,
       },
       {
-        icon: '',
-        text: 'abc',
-        title: 'abc',
         type: TextSearchResultType.Match,
+        text: 'abc',
+        start: 0,
+        end: 0,
+        lineNumber: 0,
       },
       {
-        icon: '',
-        text: 'abc',
-        title: 'abc',
         type: TextSearchResultType.Match,
+        text: 'abc',
+        start: 0,
+        end: 0,
+        lineNumber: 0,
       },
     ],
     message: 'Found 2 results in 1 file',
@@ -219,28 +224,32 @@ test('setValue - two matches in two files', async () => {
     value: 'abc',
     items: [
       {
-        icon: '',
-        text: 'file-1.txt',
-        title: '/file-1.txt',
         type: TextSearchResultType.File,
+        text: './file-1.txt',
+        start: 0,
+        end: 0,
+        lineNumber: 0,
       },
       {
-        icon: '',
-        text: 'abc',
-        title: 'abc',
         type: TextSearchResultType.Match,
+        text: 'abc',
+        start: 0,
+        end: 0,
+        lineNumber: 0,
       },
       {
-        icon: '',
-        text: 'file-2.txt',
-        title: '/file-2.txt',
         type: TextSearchResultType.File,
+        text: './file-2.txt',
+        start: 0,
+        end: 0,
+        lineNumber: 0,
       },
       {
-        icon: '',
-        text: 'abc',
-        title: 'abc',
         type: TextSearchResultType.Match,
+        text: 'abc',
+        start: 0,
+        end: 0,
+        lineNumber: 0,
       },
     ],
     message: 'Found 2 results in 2 files',
